# gateway-service/frontend-api-contract.yaml
openapi: 3.0.3
info:
  title: Telecom Dashboard Gateway Business API Contract
  version: 1.0.0
  description: |
    Frontend contract for business APIs exposed via gateway-service.
    Base URL is gateway-service (default localhost:8090).
    Event timestamps are simulation timestamps and are expected in range:
    2013-11-01T00:00:00Z .. 2013-11-07T23:59:59Z.
servers:
  - url: http://localhost:8090
...
paths:
...
  /api/alerts:
    get:
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertDto'
  /api/alerts/page:
    get:
      tags: [alerts]
      summary: Paged alerts
      parameters:
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        "200":
          description: Paged AlertDto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAlertDto'
  /api/alerts/high/stream:
    get:
      tags: [alerts]
      summary: High severity alerts SSE stream
      parameters:
        - $ref: '#/components/parameters/IntervalMs'
      responses:
        "200":
          description: text/event-stream of AlertDto[] filtered to HIGH severity
          content:
            text/event-stream:
              schema:
                type: string
  /api/alerts/critical/stream:
    get:
      tags: [alerts]
      summary: Critical severity alerts SSE stream
      parameters:
        - $ref: '#/components/parameters/IntervalMs'
      responses:
        "200":
          description: text/event-stream of AlertDto[] filtered to CRITICAL severity
          content:
            text/event-stream:
              schema:
                type: string
  /api/alerts/{id}/resolve:
    post:
      tags: [alerts]
      summary: Resolve alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Resolved
...
components:
  parameters:
...
    IntervalMs:
      name: intervalMs
      in: query
      required: false
      schema:
        type: integer
        minimum: 1000
        maximum: 300000
        default: 5000
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
      description: Spring pageable sort syntax, e.g. `timestamp,desc`
  schemas:
    AlertDto:
      description: Alert generated from congestion events.
      type: object
      properties:
        id:
          type: integer
          format: int64
        cellId:
          type: integer
        type:
          type: string
        severity:
          type: string
        message:
          type: string
        timestamp:
          type: string
          description: Simulation timestamp string in 2013-11-01..2013-11-07 window.
          example: "2013-11-03T10:00:00Z"
      required: [id, cellId, type, severity, message, timestamp]
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        offset:
          type: integer
          format: int64
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortMeta:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    PageAlertDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AlertDto'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
        size:
          type: integer
        number:
          type: integer
        sort:
          $ref: '#/components/schemas/SortMeta'
        first:
          type: boolean
        numberOfElements:
          type: integer
        empty:
          type: boolean
...